refactor the code related to streamlit UI
