+++ src/syngen/streamlit_app/handlers/__init__.py
@@ -1,0 +1,1 @@
+from syngen.streamlit_app.handlers.handlers import StreamlitHandler
